Moved `Piwik\Translate\Writer` and its subclasses to the LanguagesManager plugin
These classes are only used in that plugin, so it doesn't make sense to keep then in Core.
