+++ plugins/LanguagesManager/Test/Unit/TranslationWriter/Validate/NoScriptsTest.php
@@ -6,12 +6,12 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Tests\Unit\Translate\Validate;
+namespace Piwik\Plugins\LanguagesManager\Test\Unit\TranslationWriter\Validate;
 
-use Piwik\Translate\Validate\NoScripts;
+use Piwik\Plugins\LanguagesManager\TranslationWriter\Validate\NoScripts;
 
 /**
- * @group Translation
+ * @group LanguagesManager
  */
 class NoScriptsTest extends \PHPUnit_Framework_TestCase
 {
