+++ plugins/LanguagesManager/TranslationWriter/Filter/ByBaseTranslations.php
@@ -7,10 +7,8 @@
  *
  */
 
-namespace Piwik\Translate\Filter;
+namespace Piwik\Plugins\LanguagesManager\TranslationWriter\Filter;
 
-/**
- */
 class ByBaseTranslations extends FilterAbstract
 {
     protected $baseTranslations = array();
