+++ plugins/LanguagesManager/TranslationWriter/Filter/EmptyTranslations.php
@@ -7,10 +7,8 @@
  *
  */
 
-namespace Piwik\Translate\Filter;
+namespace Piwik\Plugins\LanguagesManager\TranslationWriter\Filter;
 
-/**
- */
 class EmptyTranslations extends FilterAbstract
 {
     /**
