+++ plugins/LanguagesManager/TranslationWriter/Filter/UnnecassaryWhitespaces.php
@@ -7,10 +7,8 @@
  *
  */
 
-namespace Piwik\Translate\Filter;
+namespace Piwik\Plugins\LanguagesManager\TranslationWriter\Filter;
 
-/**
- */
 class UnnecassaryWhitespaces extends FilterAbstract
 {
     protected $baseTranslations = array();
